+++ sdk/healthcareapis/arm-healthcareapis/src/operationsInterfaces/services.ts
@@ -14,15 +14,13 @@ import {
   ServicesListByResourceGroupOptionalParams,
   ServicesGetOptionalParams,
   ServicesGetResponse,
-  ServicesCreateOrUpdateOptionalParams,
-  ServicesCreateOrUpdateResponse,
   ServicesPatchDescription,
   ServicesUpdateOptionalParams,
   ServicesUpdateResponse,
   ServicesDeleteOptionalParams,
   CheckNameAvailabilityParameters,
   ServicesCheckNameAvailabilityOptionalParams,
-  ServicesCheckNameAvailabilityResponse
+  ServicesCheckNameAvailabilityResponse,
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -33,7 +31,7 @@ export interface Services {
    * @param options The options parameters.
    */
   list(
-    options?: ServicesListOptionalParams
+    options?: ServicesListOptionalParams,
   ): PagedAsyncIterableIterator<ServicesDescription>;
   /**
    * Get all the service instances in a resource group.
@@ -42,7 +40,7 @@ export interface Services {
    */
   listByResourceGroup(
     resourceGroupName: string,
-    options?: ServicesListByResourceGroupOptionalParams
+    options?: ServicesListByResourceGroupOptionalParams,
   ): PagedAsyncIterableIterator<ServicesDescription>;
   /**
    * Get the metadata of a service instance.
@@ -53,39 +51,8 @@ export interface Services {
   get(
     resourceGroupName: string,
     resourceName: string,
-    options?: ServicesGetOptionalParams
+    options?: ServicesGetOptionalParams,
   ): Promise<ServicesGetResponse>;
-  /**
-   * Create or update the metadata of a service instance.
-   * @param resourceGroupName The name of the resource group that contains the service instance.
-   * @param resourceName The name of the service instance.
-   * @param serviceDescription The service instance metadata.
-   * @param options The options parameters.
-   */
-  beginCreateOrUpdate(
-    resourceGroupName: string,
-    resourceName: string,
-    serviceDescription: ServicesDescription,
-    options?: ServicesCreateOrUpdateOptionalParams
-  ): Promise<
-    SimplePollerLike<
-      OperationState<ServicesCreateOrUpdateResponse>,
-      ServicesCreateOrUpdateResponse
-    >
-  >;
-  /**
-   * Create or update the metadata of a service instance.
-   * @param resourceGroupName The name of the resource group that contains the service instance.
-   * @param resourceName The name of the service instance.
-   * @param serviceDescription The service instance metadata.
-   * @param options The options parameters.
-   */
-  beginCreateOrUpdateAndWait(
-    resourceGroupName: string,
-    resourceName: string,
-    serviceDescription: ServicesDescription,
-    options?: ServicesCreateOrUpdateOptionalParams
-  ): Promise<ServicesCreateOrUpdateResponse>;
   /**
    * Update the metadata of a service instance.
    * @param resourceGroupName The name of the resource group that contains the service instance.
@@ -97,7 +64,7 @@ export interface Services {
     resourceGroupName: string,
     resourceName: string,
     servicePatchDescription: ServicesPatchDescription,
-    options?: ServicesUpdateOptionalParams
+    options?: ServicesUpdateOptionalParams,
   ): Promise<
     SimplePollerLike<
       OperationState<ServicesUpdateResponse>,
@@ -115,7 +82,7 @@ export interface Services {
     resourceGroupName: string,
     resourceName: string,
     servicePatchDescription: ServicesPatchDescription,
-    options?: ServicesUpdateOptionalParams
+    options?: ServicesUpdateOptionalParams,
   ): Promise<ServicesUpdateResponse>;
   /**
    * Delete a service instance.
@@ -126,7 +93,7 @@ export interface Services {
   beginDelete(
     resourceGroupName: string,
     resourceName: string,
-    options?: ServicesDeleteOptionalParams
+    options?: ServicesDeleteOptionalParams,
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
    * Delete a service instance.
@@ -137,7 +104,7 @@ export interface Services {
   beginDeleteAndWait(
     resourceGroupName: string,
     resourceName: string,
-    options?: ServicesDeleteOptionalParams
+    options?: ServicesDeleteOptionalParams,
   ): Promise<void>;
   /**
    * Check if a service instance name is available.
@@ -147,6 +114,6 @@ export interface Services {
    */
   checkNameAvailability(
     checkNameAvailabilityInputs: CheckNameAvailabilityParameters,
-    options?: ServicesCheckNameAvailabilityOptionalParams
+    options?: ServicesCheckNameAvailabilityOptionalParams,
   ): Promise<ServicesCheckNameAvailabilityResponse>;
 }